add story for Loader
